[V2.0.0#201912122019] Changes: - Reduced font size, not needed, I don't really know why that was that high - Tweaked some things Visual Studio was warning (not fully covered) - Added mipmap way to draw - Added some cast because Visual Studio will warn you like come on I know what I'm doing boi - Now if it doesn't know your refresh rate, it will go full automatic (worked on my 165 hz monitor pretty well) - Now it will not waste a lot of memory loading huge not needed stuff: * Max 256 MB VRAM (this one) (not precise tho) * About 250 MB system memory * Why: default font sharpness value was TOO HUGE (didn't notice because it was imported from another project I wasn't exactly aware of memory usage haha) - This is probably the best version yet (thinking about memory usage, speed and cpu/gpu efficiency) - Updated STABLE_TEST_x64.exe - This is not a huge update tho. Starting real game stuff soon!

+++ TheBlast Pack V2/TheBlast Updated V2/LSW/sound/sound.cpp
@@ -170,7 +170,7 @@ namespace LSW {
 					throw Abort::abort(__FUNCSIG__, "Failed to get PLAYMODE because there was no instance loaded.", 1);
 					return;
 				}
-				v = al_get_sample_instance_playmode(instance);
+				v = (int)al_get_sample_instance_playmode(instance);
 				break;
 			}
 		}
